Separate types for card and columns into own files
