+++ src/types/inner/card.g.ts
@@ -1,0 +1,4 @@
+export interface Card {
+  id: number;
+  title: string;
+}
