+++ src/types/inner/column.g.ts
@@ -1,0 +1,7 @@
+import { Card } from "./card.g";
+
+export interface Column {
+  id: number;
+  title?: string;
+  cards?: Card[];
+}
